Eliminating Xdebug error
Error message:  Warning: date() expects parameter 2 to be integer,
string given in
/app/public/wp-content/plugins/BE-Events-Calendar/inc/events.php on
line

Essentially, if a user creates an event and fails to set the start or
end date (doesn’t matter which), Xdebug is showing a huge error on the
backend. This eliminates that error by only running date() if, in fact,
a date has been set.