Javascript bugfix
Sometimes, when I click on a date in the calendar widget, then unclick
it, then click a second event, the first and second dates both show
their popups. This appears to have been caused by the extra logic added
for the “current” date not being specific enough and accidentally
toggling both the current date off and the other date off, then back
on. This commit fixes that minor issue, at least on my test site
(please check my work; I’m not the best in the world at javascript).