Changing the posts_per_page value to be -1
Essentially, it looks like this use case is inheriting the default
posts per page setting. I can’t envision an obvious use-case where
you’d only want up to N events showing on a calendar; it seems logical
that this is a bad default in this case.

On my test site, I ran into this in action using a bunch of recurring
events; I imagine this is simply an unintended behavior. If it’s
intentional, then that’s fine, but I don’t see an available hook to
change that behavior.